Added support for Elasticsearch 8 client
